requirements.txt
absl-py==1.1.0
aiohttp==3.9.3
aiosignal==1.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
async-timeout==4.0.3
attrs==21.4.0
autograd==1.6.2
autograd-gamma==0.5.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blosc2==2.0.0
brotlipy==0.7.0
cachetools==5.2.0
captum==0.7.0
certifi==2024.2.2
cffi==1.14.6
chardet==4.0.0
conda==4.10.3
conda-package-handling==1.7.3
contourpy==1.1.1
cryptography==3.4.7
cycler==0.11.0
Cython==3.0.8
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.15.4
fonttools==4.33.3
formulaic==1.0.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
google-auth==2.8.0
google-auth-oauthlib==0.4.6
graphlib-backport==1.0.3
grpcio==1.46.3
h5py==3.10.0
idna==2.10
imageio==2.34.1
imbalanced-learn==0.12.0
importlib-metadata==4.11.4
importlib-resources==5.8.0
interface-meta==1.3.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.8
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.17.1
jupyterlab==3.4.3
jupyterlab-language-pack-zh-CN==3.4.post1
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.3
lazy_loader==0.4
lifelines==0.27.8
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.7.5
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
notebook==6.4.12
notebook-shim==0.1.0
numexpr==2.8.6
numpy==1.22.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-python==4.9.0.80
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pip==24.0
platformdirs==3.10.0
plyfile==1.0.3
pooch==1.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
pyflann3==1.8.4.1
Pygments==2.12.0
pyOpenSSL==20.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==23.2.0
rdflib==7.0.0
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel-yaml-conda==0.15.100
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.0
setuptools==52.0.0.post20210125
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
supervisor==4.2.4
sympy==1.12.1
tables==3.8.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
threadpoolctl==2.2.0
tifffile==2023.7.10
tinycss2==1.1.1
torch==2.3.1
torch_geometric==2.5.0
torchaudio==0.12.1+cu113
torchcam==0.4.0
torchsummary==1.5.1
torchvision==0.13.1+cu113
tornado==6.1
tqdm==4.61.2
traitlets==5.3.0
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
wheel==0.36.2
widgetsnbextension==3.6.0
wrapt==1.16.0
yarl==1.9.3
zipp==3.8.0
